refactor part one
