+++ frontend/src/features/admin/components/the-contact-points-overview/ContactPointsOverview.vue
@@ -90,15 +90,15 @@ import {
   ADMIN_CONTACTPOINTS_ICON,
   ADMIN_CONTACTPOINTS_INFO_TEXT,
   ADMIN_CONTACTPOINTS_ROUTE_NAME
-} from "@/features/admin/components/contactpoints/contactPointsRoutes";
+} from "@/features/admin/components/the-contact-points-overview/contactPointsRoutes";
 import {ContactPointListItem} from "@/features/commons/types/ContactPoint";
 import {useRouter} from "vue-router/composables";
 import BackButton from "@/features/commons/components/BackButton.vue";
-import NewContactPointListItem from "@/features/admin/components/contactpoints/components/NewContactPointListItem.vue";
-import {adminContactPointLabels} from "@/features/admin/components/contactpoints/i18n";
-import NewContactPointView from "@/features/admin/components/contactpoints/components/NewContactPoint.vue";
-import NewContactPoint from "@/features/admin/components/contactpoints/components/NewContactPoint.vue";
-import EditContactPoint from "@/features/admin/components/contactpoints/components/EditContactPoint.vue";
+import NewContactPointListItem from "@/features/admin/components/the-contact-points-overview/components/NewContactPointListItem.vue";
+import {adminContactPointLabels} from "@/features/admin/components/the-contact-points-overview/i18n";
+import NewContactPointView from "@/features/admin/components/the-contact-points-overview/components/NewContactPoint.vue";
+import NewContactPoint from "@/features/admin/components/the-contact-points-overview/components/NewContactPoint.vue";
+import EditContactPoint from "@/features/admin/components/the-contact-points-overview/components/EditContactPoint.vue";
 import TheCardInitialAnlaufstellePage
   from "@/features/the-unterstuetzungsfinder/features/the-anlaufstellen/the-card-initial-anlaufstelle-page.vue";
 import {useGetContactPointListItems} from "@/features/commons/middleware/useGetContactPoints";
