+++ frontend/src/features/admin/components/the-contact-points-overview/components/EditContactPoint.vue
@@ -55,10 +55,10 @@
               <v-col cols="6">
                 <v-textarea
                   :value="writableContactPoint.description"
-                  @input="changeDescription"
                   label="Beschreibung"
                   :rules="[v => !!v || 'Beschreibung ist erforderlich', v => (v && v.length <= 2000) || 'Die Beschreibung muss weniger als 2000 Zeichen umfassen']"
                   :counter="2000"
+                  @input="changeDescription"
                 />
               </v-col>
               <v-col cols="6">
@@ -238,13 +238,13 @@ import { I18nLabel } from "@/core/core.translation";
 import { Contact, ContactPoint, ContactPointListItem, Link } from "@/features/commons/types/ContactPoint";
 import LoadingSpinner from "@/features/commons/components/LoadingSpinner.vue";
 import { marked } from "marked";
-import AddLinkDialog from "@/features/admin/components/contactpoints/components/AddLinkDialog.vue";
-import SaveUpdate from "@/features/admin/components/contactpoints/components/SaveUpdateButton.vue";
+import AddLinkDialog from "@/features/admin/components/the-contact-points-overview/components/AddLinkDialog.vue";
+import SaveUpdate from "@/features/admin/components/the-contact-points-overview/components/SaveUpdateButton.vue";
 import ErrorHandler from "@/features/commons/components/ErrorHandler.vue";
 import { useRouter } from "vue-router/composables";
-import AddContactDialog from "@/features/admin/components/contactpoints/components/AddContactDialog.vue";
-import MarkDownAlert from "@/features/admin/components/contactpoints/components/MarkDownAlert.vue";
-import DeleteButton from "@/features/admin/components/contactpoints/components/DeleteButton.vue";
+import AddContactDialog from "@/features/admin/components/the-contact-points-overview/components/AddContactDialog.vue";
+import MarkDownAlert from "@/features/admin/components/the-contact-points-overview/components/MarkDownAlert.vue";
+import DeleteButton from "@/features/admin/components/the-contact-points-overview/components/DeleteButton.vue";
 import { useGetContactPoint } from "@/features/commons/middleware/useGetContactPoints";
 export default defineComponent({
   name: "EditContactPoint",
