+++ frontend/src/features/admin/components/the-contact-points-overview/components/NewContactPointListItem.vue
