+++ frontend/src/features/admin/components/the-contact-points-overview/components/SaveNewButton.vue
@@ -25,7 +25,7 @@
 <script lang="ts">
 import { defineComponent, ref } from "vue";
 import { ContactPoint } from "@/features/commons/types/ContactPoint";
-import { useCreateNewContactPoint } from "@/features/admin/components/contactpoints/middelware/useContactPoints";
+import { useCreateNewContactPoint } from "@/features/admin/components/the-contact-points-overview/middelware/useContactPoints";
 import { useRouter } from "vue-router/composables";
 
 export default defineComponent({
