+++ frontend/src/features/admin/components/the-contact-points-overview/i18n.ts
