+++ frontend/src/features/admin/components/the-contact-points-overview/middelware/ListUtils.ts
